scripts/requirements.txt
python-dotenv==1.0.1
supabase==2.6.0
requests==2.32.3

